requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.35.3
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
PyMySQL==1.1.1
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
